스프라이트 애니메이션 구조 구축중.

+++ swmodule/header/SWSprite.h
@@ -10,7 +10,7 @@ class SWSprite : public SWObject
 
 public:
 
-	SWSprite( SWHardRef<SWTexture> texture, float x, float y, float width, float height );
+	SWSprite( SWHardRef<SWTexture> texture, int x, int y, int width, int height );
 	~SWSprite();
 
 	SWTexture* getTexture() const;
